I have do so many hacky things to get the character builder running.
My childrens children will pay for the sins I've committed for this code. But it was quick and easy to get the functionality somewhat in place to see if I had the right idea.

It seems I'm on the right path. The character builder is working enough that I can create and save characters of all kinds. This code needs cleanup beyond imagination but I'll get there.